Establish the property->object->context chain of reference.
Change addProperty to a method. Each property now holds a ptr to the object
it belongs to. Add addObject. Now each object has a ptr to the echonet lite
context. This should be enough to prevent the use of one global context
structure but keep the API reasonably simple going forward.